Store version in a separate module
