+++ .gitignore
@@ -1,5 +1,6 @@
 # Byte-compiled / optimized / DLL files
 __pycache__/
+musicbrainzngs.egg-info/
 *.py[cod]
 *$py.class
