+++ musicbrainzngs/__init__.py
@@ -1,2 +1,3 @@
+from musicbrainzngs.version import _version
 from musicbrainzngs.musicbrainz import *
 from musicbrainzngs.caa import *
