+++ musicbrainzngs/version.py
@@ -1,0 +1,1 @@
+_version = "0.7.1"
